fix: split up the store layout
This changes the layout of the file system store to split content up
by distribution as well as the prefix directories (e.g. `2023`)
underneath that.

Having too many files in a single directory can become a performance
issue, so it is necessary to split this up. Also, it was possible to
for documents with the same name to overlap in different distributions.

Instead of coming up with a custom logic, this follows the idea of CSAF,
having a prefix directory by year.

NOTE: This changes the filesystem layout of the stored files. However,
that layout was never considered an API, so it is not considered a
breaking change. But one should be aware of that.