fix: drop the error prone DNS error check for a DNS pre-flight check
There is no clear indication if a HTTP request failed due to a
"host not found" condition. So we do a DNS pre-flight check to see if
the expected DNS name resolves. If it does, we do a simple HTTP request.

If it does not resolve, we consider this step to return "no content".