feat: add scoop command for the SBOM cli
This allows using a more local approach then the send command. Locating
local files and sending them to the target.

The difference to the send command is, that it doesn't require any
specific filesystem layout. And it also supports cleaning up processed
(or failed) files. So that it can be re-run on the same directory.